# openapi.yaml
openapi: 3.0.4
info:
  description: |
    The **Tag Management System API** provides RESTful access to all tag-related operations—creation, updates, deletion, linking, and querying of tags and their relationships.
    It also supports metadata assignment for tagged entities such as images, documents, or video stills.

    ## Principles
    - **RESTful architecture** — resource-based URIs and standard HTTP verbs (GET, POST, PATCH, DELETE)
    - **JSON payloads** — all requests and responses use JSON
    - **Frontend-controlled authentication** — only frontend clients (CLI, UI) interact with the API through validated tokens (except for development/testing)
  title: Tag Management System API
  version: 0.1.1
servers:
- description: Local development
  url: http://localhost:8000/api/v1
...
paths:
...
  /ui/groups:
    get:
      description: Retrieve all reusable UI group definitions that segment layouts
        into labeled sections.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_ui_groups_get_200_response"
          description: List of UI groups
      summary: List all UI groups
      tags:
      - ui:groups
    post:
      description: Create a labeled section definition that can be reused across UI
        layouts.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UiGroupInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiGroup"
          description: Group created
      summary: Create a new UI group
      tags:
      - ui:groups
  /ui/groups/{id}:
    delete:
      description: Permanently remove a UI group. Layouts using this group must be
        updated accordingly.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_tags__id__delete_200_response"
          description: Group deleted
      summary: Delete a specific UI group
      tags:
      - ui:groups
    get:
      description: Retrieve details for a UI group including its name and associated
        layouts.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiGroup"
          description: Group found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Group not found
      summary: Retrieve a specific UI group
      tags:
      - ui:groups
    patch:
      description: Update the name or metadata for a UI group.
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UiGroupInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UiGroup"
          description: Group updated
      summary: Update a specific UI group
      tags:
      - ui:groups
...
components:
...
  schemas:
    ErrorResponse:
      example:
        code: 404
        details: "{}"
        message: Resource not found
      properties:
        code:
          example: 404
          title: code
          type: integer
        message:
          example: Resource not found
          title: message
          type: string
        details:
          title: details
          type: object
      required:
      - code
      - message
      title: ErrorResponse
      type: object
...
    UiGroup:
      description: A labeled section within a UI layout used to organize fields under
        a specific theme or purpose.
      example:
        updated_at: 2000-01-23T04:56:07.000+00:00
        name: name
        created_at: 2000-01-23T04:56:07.000+00:00
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: id
          type: string
        name:
          title: name
          type: string
        created_at:
          format: date-time
          title: created_at
          type: string
        updated_at:
          format: date-time
          title: updated_at
          type: string
      title: UiGroup
      type: object
    UiGroupInput:
      example:
        name: name
      properties:
        name:
          title: name
          type: string
      required:
      - name
      title: UiGroupInput
      type: object
...
    _tags__id__delete_200_response:
      example:
        status: deleted
      properties:
        status:
          example: deleted
          title: status
          type: string
      title: _tags__id__delete_200_response
      type: object
...
    _ui_groups_get_200_response:
      example:
        ui_groups:
        - updated_at: 2000-01-23T04:56:07.000+00:00
          name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        - updated_at: 2000-01-23T04:56:07.000+00:00
          name: name
          created_at: 2000-01-23T04:56:07.000+00:00
          id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        ui_groups:
          items:
            $ref: "#/components/schemas/UiGroup"
          title: ui_groups
          type: array
      title: _ui_groups_get_200_response
      type: object
...